src/test-server/requirements.txt
click==8.1.2
Flask==2.1.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
PyYAML==6.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
zipp==3.8.0
